# Simplify SHA256x8 for the compiler