--- a/ref/sha256.h
+++ b/ref/sha256.h
@@ -21,7 +21,7 @@ void sha256(uint8_t *out, const uint8_t *in, size_t inlen);
 void mgf1(unsigned char *out, unsigned long outlen,
           const unsigned char *in, unsigned long inlen);
 
-uint8_t state_seeded[40];
+extern uint8_t state_seeded[40];
 
 void seed_state(const unsigned char *pub_seed);