--- a/sha256-avx2/sha256avx.h
+++ b/sha256-avx2/sha256avx.h
@@ -73,16 +73,6 @@ typedef struct SHA256state {
 
 void transpose(u256 s[8]);
 void sha256_init8x(sha256ctx *ctx);
-void sha256_update8x(sha256ctx *ctx, 
-                     const unsigned char *d0,
-                     const unsigned char *d1,
-                     const unsigned char *d2,
-                     const unsigned char *d3,
-                     const unsigned char *d4,
-                     const unsigned char *d5,
-                     const unsigned char *d6,
-                     const unsigned char *d7,
-                     unsigned long long len);
 void sha256_final8x(sha256ctx *ctx,
                     unsigned char *out0,
                     unsigned char *out1,
@@ -93,7 +83,14 @@ void sha256_final8x(sha256ctx *ctx,
                     unsigned char *out6,
                     unsigned char *out7);
 
-void sha256_transform8x(sha256ctx *ctx, const unsigned char *data);
-
+void sha256_transform8x(sha256ctx *ctx,
+        const unsigned char *data0,
+        const unsigned char *data1,
+        const unsigned char *data2,
+        const unsigned char *data3,
+        const unsigned char *data4,
+        const unsigned char *data5,
+        const unsigned char *data6,
+        const unsigned char *data7);
 
 #endif