--- a/sha256-avx2/thash_sha256_simplex8.c
+++ b/sha256-avx2/thash_sha256_simplex8.c
@@ -33,11 +33,8 @@ void thashx8(unsigned char *out0,
     unsigned char bufx8[8*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N)];
     unsigned char outbufx8[8*SPX_SHA256_OUTPUT_BYTES];
     unsigned int i;
-    sha256ctx ctx;
 
     (void)pub_seed; /* Suppress an 'unused parameter' warning. */
-
-    sha256_init_frombytes_x8(&ctx, state_seeded, 512);
 
     for (i = 0; i < 8; i++) {
         memcpy(bufx8 + i*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
@@ -61,26 +58,31 @@ void thashx8(unsigned char *out0,
     memcpy(bufx8 + SPX_SHA256_ADDR_BYTES +
         7*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N), in7, inblocks * SPX_N);
 
-    sha256_update8x(&ctx,
-                    bufx8 + 0*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 1*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 2*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 3*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 4*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 5*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 6*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    bufx8 + 7*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
-                    SPX_SHA256_ADDR_BYTES + inblocks*SPX_N);
+    sha256x8_seeded(
+        /* out */
+        outbufx8 + 0*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 1*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 2*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 3*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 4*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 5*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 6*SPX_SHA256_OUTPUT_BYTES,
+        outbufx8 + 7*SPX_SHA256_OUTPUT_BYTES,
 
-    sha256_final8x(&ctx,
-                   outbufx8 + 0*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 1*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 2*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 3*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 4*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 5*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 6*SPX_SHA256_OUTPUT_BYTES,
-                   outbufx8 + 7*SPX_SHA256_OUTPUT_BYTES);
+        /* seed */
+        state_seeded, 512,
+
+        /* in */
+        bufx8 + 0*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 1*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 2*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 3*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 4*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 5*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 6*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        bufx8 + 7*(SPX_SHA256_ADDR_BYTES + inblocks*SPX_N),
+        SPX_SHA256_ADDR_BYTES + inblocks*SPX_N /* len */
+    );
 
     memcpy(out0, outbufx8 + 0*SPX_SHA256_OUTPUT_BYTES, SPX_N);
     memcpy(out1, outbufx8 + 1*SPX_SHA256_OUTPUT_BYTES, SPX_N);